Can add and delete oscillators

+++ src/playground/base_audio_element.cpp
@@ -3,6 +3,6 @@
 namespace Playground
 {
 
-	unsigned int BaseAudioElement::_counter = 0;
+	size_t BaseAudioElement::_counter = 0;
 
 } // namespace Playground
